Don't discard unwanted messages
Right now we process messages with a queue, and discard any that don't
match the id that we're waiting for. This means that messages can't be
processed out of order, which can happen with a multithreaded
architecture. I've refactored it to use a synchronized hash rather
than a queue, which allows us to retrieve messages in any order.