Only continue intercepted request if it passes black & white lists and the black/whitelist callback is the last callback
Previously, if the request didn't match the blacklist,
or matched the whitelist, it'd be continued, without regard
for if there were other callbacks. This prohibited those
callbacks from aborting the request or responding to it.

This also permits blacklists and whitelists to be combined.
Before, if both a white and blacklist were configured,
the whitelist would be ignored. Now, if both are configured,
only requests that do not match the blacklist and match
the whitelist are permitted.