fix: correctly open the devtools
The original debug url was deprecated apparently (I didn't confirm that;
    a call to `page.driver.debug` just opened a blank page).

Instead there is an endpoint now that allows to parse out a remote
devtools frontend path (although they incorrectly call it a url).

So instead of generating a fixed debug_url now we make a request to said
endpoint and parse the actual (dynamic; contains a uuid) devtools path
and build a url from it.