Docs: Update url allow/block instructions (#265)
Currently, if you use a string to mark an allowed or blocked domain/url
you get a noisy deprecation warning about the use of strings for these
parameters. This change updates the instructions to specify the use of
regexes instead which conforms to what Ferrum's code seems to expect
(`pattern`) and silences the warning.