Ensure node has focus before setting value
Currently the javascript used to set a field value will trigger
a "focus" event prior to updating a field however this does not
set the input as focused so any use of `document.activeElement`
will return the body element.

To fix this a call to `node.focus()` has been added before the
focus event to ensure `document.activeElement` will respond
correctly.